Fixing warm start and updating reference results for MCT

+++ tests/test_utils.py
@@ -242,7 +242,8 @@ class ETRS89TestCase(object):
     def teardown_method(self):
         settings = LisSettings.instance()
         output_dir = settings.output_dir
-        shutil.rmtree(output_dir)
+        if os.path.exists(output_dir):
+            shutil.rmtree(output_dir)
 
     @classmethod
     def compare_reference(cls, variable='dis', check='map', step_length='86400'):
@@ -324,7 +325,8 @@ class MCTTestCase(object):
     def teardown_method(self):
         settings = LisSettings.instance()
         output_dir = settings.output_dir
-        shutil.rmtree(output_dir)
+        if os.path.exists(output_dir):
+            shutil.rmtree(output_dir)
 
     @classmethod
     def compare_reference(cls, variable='dis', check='map', step_length='86400'):
